Changed get_proposal_situation to model save
